pyproject.toml
[tool.poetry]
name = "nonebot-plugin-morning"
version = "0.3.0a4"
description = "Good morning & good night!"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_morning"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b3"
nonebot-adapter-onebot = "^2.1.1"
nonebot-plugin-apscheduler = "^0.1.3"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true
